requirements.txt
Pillow>=10.0.0
pyautogui>=0.9.53
pytesseract>=0.3.10
pystray>=0.19.5
requests>=2.31.0
keyboard>=0.13.5
PyPDF2>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0